+++ fj-doc-base-json/src/main/java/org/fugerit/java/doc/json/typehandler/DocTypeHandlerCoreJSON.java
@@ -18,7 +18,7 @@ public class DocTypeHandlerCoreJSON extends DocTypeHandlerDefault {
 
     public static final DocTypeHandler HANDLER_UTF8 = new DocTypeHandlerCoreJSON( StandardCharsets.UTF_8 );
 
-    public static final String TYPE = DocConfig.TYPE_XML;
+    public static final String TYPE = DocConfig.TYPE_JSON;
 
     public static final String MODULE = "doc-core";