+++ CHANGELOG.md
@@ -11,7 +11,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - type handlers for XML (DocTypeHandlerCoreXML and DocTypeHandlerCoreXMLUTF8)
 - type handlers for JSON (DocTypeHandlerCoreJSON and DocTypeHandlerCoreJSONUTF8)
-- - type handlers for JSON (DocTypeHandlerCoreYAML and DocTypeHandlerCoreYAMLUTF8)
+- type handlers for JSON (DocTypeHandlerCoreYAML and DocTypeHandlerCoreYAMLUTF8)
+
+### Fixed
+
+- org.fugerit.java.doc.base.config.DocTypeHandlerXML renders the source document in the source format (i.e. JSON or YAML) <https://github.com/fugerit-org/fj-doc/issues/519>
 
 ## [8.16.5] - 2025-09-29